refactor(runtime): extract relay auth/health/info cache with PWA adapters
